Add IMU covariance propagation
