+++ pypose/__init__.py
@@ -6,5 +6,5 @@ from .liegroup.group import identity_Sim3, identity_sim3, identity_RxSO3, identi
 from .liegroup.group import Exp, Log, Inv, Mul, Retr, Act, Adj, AdjT, Jinv
 from .liegroup.group import SO3_type, so3_type, SE3_type, se3_type
 from .liegroup.group import Sim3_type, sim3_type, RxSO3_type, rxso3_type
-from .liegroup.group import mat2SO3, euler2SO3
+from .liegroup.group import mat2SO3, euler2SO3, vec2skew
 from .module import IMUPreintegrator
