+++ pypose/liegroup/group/__init__.py
@@ -1,4 +1,5 @@
 __all__ = ['groups']
+from .basics import vec2skew
 from .groups import LieGroup, Parameter
 from .groups import SO3_type, so3_type, SE3_type, se3_type
 from .groups import Sim3_type, sim3_type, RxSO3_type, rxso3_type
@@ -8,4 +9,4 @@ from .utils import identity_like, identity_SO3, identity_so3, identity_SE3, iden
 from .utils import identity_Sim3, identity_sim3, identity_RxSO3, identity_rxso3
 from .utils import SO3, so3, SE3, se3, Sim3, sim3, RxSO3, rxso3
 from .utils import Exp, Log, Inv, Mul, Retr, Act, Adj, AdjT, Jinv
-from .basics import mat2SO3, euler2SO3
+from .convert import mat2SO3, euler2SO3
