+++ pypose/liegroup/group/convert.py
@@ -1,0 +1,100 @@
+import torch
+from .utils import SO3, so3
+
+
+def mat2SO3(rotation_matrix, **kwargs):
+    """Convert 3x3 or 3x4 rotation matrix to SO3 group (by quaternion)
+
+    Args:
+        rotation_matrix (Tensor): the rotation matrix to convert.
+
+    Return:
+        Tensor: the SO3 group in quaternion
+
+    Shape:
+        - Input: :math:`(*, 3, 3)` or '(*, 3, 4)'
+        - Output: :math:`(*, 4)`
+
+    Example:
+        >>> input = torch.rand(4, 3, 3)  # Nx3x3
+        >>> output = pp.mat2SO3(input)  # Nx4
+    """
+    if not torch.is_tensor(rotation_matrix):
+        rotation_matrix = torch.tensor(rotation_matrix)
+
+    if len(rotation_matrix.shape) > 3:
+        raise ValueError(
+            "Input size must be a three dimensional tensor. Got {}".format(
+                rotation_matrix.shape))
+    if not (rotation_matrix.shape[-2:] == (3, 3) or rotation_matrix.shape[-2:] == (3, 4)):
+        raise ValueError(
+            "Input size must be a * x 3 x 4 or * x 3 x 3  tensor. Got {}".format(
+                rotation_matrix.shape))
+    shape = rotation_matrix.shape
+    rotation_matrix = rotation_matrix.view(-1, shape[-2], shape[-1])
+    rmat_t = torch.transpose(rotation_matrix, 1, 2)
+
+    mask_d2 = rmat_t[:, 2, 2] < 1e-6
+
+    mask_d0_d1 = rmat_t[:, 0, 0] > rmat_t[:, 1, 1]
+    mask_d0_nd1 = rmat_t[:, 0, 0] < -rmat_t[:, 1, 1]
+
+    t0 = 1 + rmat_t[:, 0, 0] - rmat_t[:, 1, 1] - rmat_t[:, 2, 2]
+    q0 = torch.stack([rmat_t[:, 1, 2] - rmat_t[:, 2, 1],
+                      t0, rmat_t[:, 0, 1] + rmat_t[:, 1, 0],
+                      rmat_t[:, 2, 0] + rmat_t[:, 0, 2]], -1)
+    t0_rep = t0.repeat(4, 1).t()
+
+    t1 = 1 - rmat_t[:, 0, 0] + rmat_t[:, 1, 1] - rmat_t[:, 2, 2]
+    q1 = torch.stack([rmat_t[:, 2, 0] - rmat_t[:, 0, 2],
+                      rmat_t[:, 0, 1] + rmat_t[:, 1, 0],
+                      t1, rmat_t[:, 1, 2] + rmat_t[:, 2, 1]], -1)
+    t1_rep = t1.repeat(4, 1).t()
+
+    t2 = 1 - rmat_t[:, 0, 0] - rmat_t[:, 1, 1] + rmat_t[:, 2, 2]
+    q2 = torch.stack([rmat_t[:, 0, 1] - rmat_t[:, 1, 0],
+                      rmat_t[:, 2, 0] + rmat_t[:, 0, 2],
+                      rmat_t[:, 1, 2] + rmat_t[:, 2, 1], t2], -1)
+    t2_rep = t2.repeat(4, 1).t()
+
+    t3 = 1 + rmat_t[:, 0, 0] + rmat_t[:, 1, 1] + rmat_t[:, 2, 2]
+    q3 = torch.stack([t3, rmat_t[:, 1, 2] - rmat_t[:, 2, 1],
+                      rmat_t[:, 2, 0] - rmat_t[:, 0, 2],
+                      rmat_t[:, 0, 1] - rmat_t[:, 1, 0]], -1)
+    t3_rep = t3.repeat(4, 1).t()
+
+    mask_c0 = mask_d2 * mask_d0_d1
+    mask_c1 = mask_d2 * ~mask_d0_d1
+    mask_c2 = ~mask_d2 * mask_d0_nd1
+    mask_c3 = ~mask_d2 * ~mask_d0_nd1
+    mask_c0 = mask_c0.view(-1, 1).type_as(q0)
+    mask_c1 = mask_c1.view(-1, 1).type_as(q1)
+    mask_c2 = mask_c2.view(-1, 1).type_as(q2)
+    mask_c3 = mask_c3.view(-1, 1).type_as(q3)
+
+    q = q0 * mask_c0 + q1 * mask_c1 + q2 * mask_c2 + q3 * mask_c3
+    q /= 2*torch.sqrt(t0_rep * mask_c0 + t1_rep * mask_c1 +  # noqa
+                    t2_rep * mask_c2 + t3_rep * mask_c3)  # noqa
+
+    q = q.view(shape[:-2]+(4,))
+    q = q.index_select(-1, torch.tensor([1,2,3,0])) # wxyz -> xyzw
+
+    return SO3(q, **kwargs)
+
+
+def euler2SO3(euler:torch.Tensor, **kwargs):
+    # euler: (*, 3) in sequence of roll, pitch, yaw
+    if not torch.is_tensor(euler):
+        euler = torch.tensor(euler)
+    assert euler.shape[-1] == 3
+    shape, euler = euler.shape, euler.view(-1, 3)
+    roll, pitch, yaw = euler[:,0], euler[:,1], euler[:,2]
+    cy, sy = (yaw * 0.5).cos(), (yaw * 0.5).sin()
+    cp, sp = (pitch * 0.5).cos(), (pitch * 0.5).sin()
+    cr, sr = (roll * 0.5).cos(), (roll * 0.5).sin()
+
+    q = torch.stack([sr * cp * cy - cr * sp * sy,
+                     cr * sp * cy + sr * cp * sy,
+                     cr * cp * sy - sr * sp * cy,
+                     cr * cp * cy + sr * sp * sy], dim=-1)
+    return SO3(q, **kwargs).gview(*shape[:-1])
